managemate_bacend/requirement.txt
Django==2.1.5
django-cors-headers==2.5.2
django-filter==2.1.0
django-postgres-unlimited-varchar==1.0.1
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
Markdown==3.0.1
Pillow==5.4.1
postgres==2.2.2
psycopg2-binary==2.7.7
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
